requirements.txt
Flask
pyserial
pynmea2
colorama